Optimize day 14 2024 by looping in bigger steps
The x-axis loops every width-steps, so we only need to find the first
second where the robots are already on the correct x-axis (counting the
borders of the tree), then we only need to loop in these steps until
also the y-axis is correct.

There is probably also a mathematical solution by doing the same with
the y-axis and then calculate it with the size and these 2 offsets, but
my brain can't think about this at the moment 🤯

This is already a really big performance improvements, so "good enough".